Let room generate kotlin

+++ app/src/main/java/com/hippo/ehviewer/EhDB.kt
@@ -243,7 +243,9 @@ object EhDB {
     @Synchronized
     fun deleteQuickSearch(quickSearch: QuickSearch?) {
         val dao = db.quickSearchDao()
-        dao.delete(quickSearch)
+        if (quickSearch != null) {
+            dao.delete(quickSearch)
+        }
     }
 
     @Synchronized
